Allocation refactor (#4113)
Stacked on #4125 

This PR reduces some of the duplicated logic between the allocation and
indexing lowering passes. Both passes need to infer correct iter domains
to allocate. In this PR, the inference code of the indexing pass is
moved to the allocation pass, and the result of the inference done in
the allocation pass is reused in the indexing pass.

This is purely a refactoring pass. It should not have any impact to the
final results. However, because the inference logic used in the indexing
pass is different from the one currently used in the allocation pass,
there are some differences in generated code:

- They are mostly just like `i1 * i2` instead of `i2 * i1`, so
mathematically equivalent.
- In one case, the new version allocates a smaller shared memory region
because of the combination of this transpose
[logic](https://github.com/NVIDIA/Fuser/blob/main/csrc/id_model/indexing.cpp#L560)
and broadcast. I verified that is the correct allocation size.
- It seems that the new logic can find more aliasing opportunities.
