+++ csrc/device_lower/pass/index.cpp
@@ -8,7 +8,6 @@
 #include <device_lower/analysis/index_compute.h>
 #include <device_lower/analysis/tma.h>
 #include <device_lower/lower2device.h>
-#include <device_lower/utils.h>
 #include <id_model/schedule.h>
 #include <index_compute.h>
 #include <ir/iostream.h>
@@ -30,11 +29,6 @@ namespace nvfuser {
 std::vector<Expr*> IndexLowering::getIndexedExprs(
     std::vector<Expr*> incoming_exprs) {
   FUSER_PERF_SCOPE("GpuLower::Lower::IndexLowering::getIndexedExprs");
-  // Traverse the exprs and setup allocation domains before
-  // generating indices.
-  if (GpuLower::current()->isTensorIndexerEnabled()) {
-    GpuLower::current()->tensorIndexer().setupAllocationDomains(incoming_exprs);
-  }
   IndexLowering il;
   il.generate(incoming_exprs);
   return il.lowered_exprs_;
